LoadGen: Create test_settings.h; Clean up headers.
