+++ loadgen/loadgen.cc
@@ -15,6 +15,7 @@
 #include "query_sample.h"
 #include "query_sample_library.h"
 #include "system_under_test.h"
+#include "test_settings.h"
 #include "utils.h"
 #include "version.h"
